connectionCount 타입 변환

+++ backend/src/main/java/com/whyweclimb/backend/entity/SimultaneousConnection.java
@@ -19,7 +19,7 @@ public class SimultaneousConnection {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer simultaneousConnectionSeq;
     @Column
-    private Integer connectionCount;
+    private Long connectionCount;
     @Column
     private LocalDate connectionDate;
 }
